auto refresh delete works
