+++ app/build.gradle
@@ -86,6 +86,10 @@ dependencies {
     implementation "androidx.recyclerview:recyclerview-selection:1.1.0-rc03"
 
 
+    //lottie
+    implementation 'com.airbnb.android:lottie:3.0.7'
+
+
 
     //room
     def room_version = "2.2.6"
