+++ app/src/main/java/com/example/moulinapplication/ui/dashboard/DashboardFragment.kt
@@ -1,43 +1,30 @@
 package com.example.moulinapplication.ui.dashboard
 
-import android.app.Dialog
-import android.opengl.GLES30
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.LinearLayout
-import android.widget.TextView
-import android.widget.Toast
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.moulinapplication.R
 import com.example.moulinapplication.databinding.FragmentDashboardBinding
-import com.example.moulinapplication.databinding.FragmentPopUpBinding
 import com.example.moulinapplication.model.Joke
 import com.example.moulinapplication.network.RetrofitBuilder
 import com.example.moulinapplication.repositories.JokeRepo
 import com.example.moulinapplication.roomdb.JokeDao
 import com.example.moulinapplication.roomdb.RoomDB
-import com.example.moulinapplication.ui.home.AddJokeFragmentViewModel
-import com.example.moulinapplication.ui.home.AddJokeFragmentViewModelFactory
 import com.example.moulinapplication.ui.popup.PopUpFragment
-import retrofit2.Retrofit
-import timber.log.Timber
 
 class DashboardFragment : Fragment() {
 
     private lateinit var dashboardViewModel: DashboardViewModel
     lateinit var binding: FragmentDashboardBinding
-    lateinit var adapter : JokeRecyclerViewAdapter
+    lateinit var adapter: JokeRecyclerViewAdapter
 
     override fun onCreateView(
-            inflater: LayoutInflater,
-            container: ViewGroup?,
-            savedInstanceState: Bundle?
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
     ): View? {
 
 
@@ -48,32 +35,32 @@ class DashboardFragment : Fragment() {
         dashboardViewModel = ViewModelProvider(this, factory).get(DashboardViewModel::class.java)
 
 
-
         //init binding
         binding = FragmentDashboardBinding.inflate(inflater, container, false)
+        binding.lifecycleOwner = this
 
 
-
-        //recycler view vullen met de favourite jokes
-        dashboardViewModel.jokes.observe(viewLifecycleOwner, {
-            binding.recyclerview.layoutManager = LinearLayoutManager(this.requireContext())
-            binding.lifecycleOwner = this
-            adapter = JokeRecyclerViewAdapter(it, { selected: Joke -> jokeIsClicked(selected) })
-            adapter.notifyDataSetChanged()
-            binding.recyclerview.adapter = adapter
-        })
+        //init recyclerview
+        initrecyclerview()
 
 
         return binding.root
     }
 
 
-    fun jokeIsClicked(joke: Joke) {
-        //show pop up dialog
-        val dialog = PopUpFragment(joke)
-        getFragmentManager()?.let { dialog.show(it, "popUpDialog") }
-
+    fun initrecyclerview(){
+        dashboardViewModel.jokes.observe(viewLifecycleOwner, {
+            adapter = JokeRecyclerViewAdapter(it,{ selected: Joke -> jokeIsClicked(selected) })
+            binding.recyclerview.layoutManager = LinearLayoutManager(this.requireContext())
+            binding.recyclerview.adapter = adapter
+        })
     }
 
 
+    fun jokeIsClicked(joke: Joke) {
+        //show pop up dialog
+        val dialog = PopUpFragment(joke,dashboardViewModel,adapter)
+        getFragmentManager()?.let { dialog.show(it, "popUpDialog") }
+
+    }
 }
