+++ app/src/main/java/com/example/moulinapplication/ui/home/HomeViewModel.kt
@@ -36,7 +36,4 @@ class HomeViewModel(private val jokeRepo: JokeRepo): ViewModel() {
         }
     }
 
-
-
-
 }
