+++ app/src/main/java/com/example/moulinapplication/ui/popup/PopUpFragment.kt
@@ -6,10 +6,13 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.fragment.app.DialogFragment
+import androidx.fragment.app.FragmentPagerAdapter
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.Navigation
 import androidx.navigation.findNavController
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.moulinapplication.R
 
 import com.example.moulinapplication.databinding.FragmentPopUpBinding
@@ -17,37 +20,29 @@ import com.example.moulinapplication.model.Joke
 import com.example.moulinapplication.network.RetrofitBuilder
 import com.example.moulinapplication.repositories.JokeRepo
 import com.example.moulinapplication.roomdb.RoomDB
+import com.example.moulinapplication.ui.dashboard.DashboardViewModel
+import com.example.moulinapplication.ui.dashboard.JokeRecyclerViewAdapter
 import timber.log.Timber
 
 
-class PopUpFragment(private val selectedJoke : Joke) : DialogFragment() {
+class PopUpFragment(private val selectedJoke : Joke,val dashBoardViewmodel: DashboardViewModel,val adapter: JokeRecyclerViewAdapter) : DialogFragment() {
 
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
-
-        //init viewmodel
-        val dao = RoomDB.getInstance(this.requireContext()).JokeDAO
-        val jokerepo = JokeRepo(dao,RetrofitBuilder.jokeservice)
-        val factory = PopUpViewModelFactory(jokerepo)
-        val popupViewModel : PopUpFragmentViewModel = ViewModelProvider(this,factory).get(PopUpFragmentViewModel::class.java)
 
         //init binding
         val binding = FragmentPopUpBinding.inflate(inflater,container,false)
 
         //edit clicked
         binding.editjoke.setOnClickListener{
-            //TODO
-            //popupViewModel.deleteJoke(it)
-            //popupViewModel.editJoke(selectedJoke)
-            //this.dismiss()
         }
 
 
         //delete clicked
         binding.deletejoke.setOnClickListener {
-            //TODO
-            popupViewModel.deleteJoke(selectedJoke)
+            dashBoardViewmodel.deleteJoke(selectedJoke)
+            Toast.makeText(this.requireContext(),"Joke deleted successfully", Toast.LENGTH_LONG)
             this.dismiss()
         }
